requirements.txt
# ---- Core Python runtime ----
python==3.10.*

# ---- Deep Learning Framework ----
torch==2.4.*+cu121
torchvision==0.19.*+cu121
torchaudio==2.4.*+cu121
# 上述 +cu121 对应 CUDA 12.1 wheels，可在 12.2 驱动上正常运行

# ---- HuggingFace Transformers ----
transformers>=4.44.0

# ---- Utilities ----
pillow>=10.3.0
requests>=2.31.0

tqdm
numpy